04_code/ai-sdk-rag-starter/simulation/requirements.txt
openai>=1.68.0
pydantic>=2.8.0
python-dotenv>=1.0.0
requests>=2.31.0
colorama>=0.4.6